Refactor join dialog component
