+++ src/lib/components/GroupSelection.svelte
@@ -1,0 +1,156 @@
+<script lang="ts">
+    import { goto } from "$app/navigation";
+    import { joinDialogOpen, selectedCourse } from "$lib/client/state.svelte";
+    import * as AlertDialog from "$lib/components/ui/alert-dialog";
+    import { ChevronLeft, Icon, Plus } from "svelte-hero-icons";
+    import { toast } from "svelte-sonner";
+    import Button from "./ui/button/button.svelte";
+    import { onMount } from "svelte";
+
+    type Group = {
+        groupId: number;
+        groupName: string;
+        members: {
+            netid: string;
+            displayname: string;
+        }[];
+    };
+
+    let isLoading = $state(false);
+    let availableGroups: Group[] = $state([]);
+    let alertDialogOpen = $state(false);
+
+    async function loadGroups() {
+        isLoading = true;
+        try {
+            const res = await fetch(
+                `/api/groups?courseId=${selectedCourse.value!.id}`
+            );
+            const data = await res.json();
+            availableGroups = data;
+        } finally {
+            isLoading = false;
+        }
+    }
+
+    const createNewGroup = async () => {
+        if (!selectedCourse.value) {
+            alertDialogOpen = false;
+            return;
+        }
+
+        const res = await fetch("/api/new-group", {
+            method: "POST",
+            headers: { "Content-Type": "application/json" },
+            body: JSON.stringify({ courseId: selectedCourse.value.id })
+        });
+        const newGroupId = (await res.json()).group;
+
+        goto(`/group/${newGroupId}`);
+        toast.success(`Created a new group for ${selectedCourse.value.code}!`);
+        alertDialogOpen = false;
+        joinDialogOpen.value = false;
+    };
+
+    const joinGroup = async (group: Group) => {
+        await fetch("/api/join-group", {
+            method: "POST",
+            headers: { "Content-Type": "application/json" },
+            body: JSON.stringify({ groupId: group.groupId })
+        });
+        joinDialogOpen.value = false;
+        goto(`/group/${group.groupId}`);
+        toast.success(`Joined group ${group.groupName}!`);
+    };
+
+    onMount(() => {
+        loadGroups();
+    });
+</script>
+
+<div class="my-4">
+    <h2 class="flex items-center gap-2 text-xl font-semibold">
+        <Button
+            size="icon"
+            variant="ghost"
+            onclick={() => {
+                selectedCourse.value = null;
+            }}>
+            <Icon src={ChevronLeft} size="32" />
+        </Button>
+        <span>Groups for {selectedCourse.value!.code}</span>
+    </h2>
+    <p class="text-sm text-slate-500">
+        Select a group to join or create a new one.
+    </p>
+</div>
+
+{#if isLoading}
+    <div class="std-flex mt-4">Loading groups...</div>
+{:else if availableGroups.length === 0}
+    <div class="std-flex flex-1 text-slate-500">
+        No groups found. Create a new one?
+    </div>
+{:else}
+    <div class="flex flex-1 flex-col overflow-y-auto border-t border-input">
+        {#each availableGroups as group (group.groupId)}
+            <button onclick={() => joinGroup(group)} class="card">
+                <div class="p-2">
+                    <p class="text-sm font-semibold">
+                        Group: {group.groupName}
+                    </p>
+                    <p class="text-xs text-slate-500">
+                        <span class="font-semibold">
+                            {group.members.length} Members:
+                        </span>
+                        <span>
+                            {group.members.map((x) => x.displayname).join(", ")}
+                        </span>
+                    </p>
+                </div>
+            </button>
+        {:else}
+            <div class="flex-1 std-flex text-slate-500">
+                You are already in all available groups.
+            </div>
+        {/each}
+    </div>
+{/if}
+
+<Button
+    onclick={() => {
+        if (availableGroups.length === 0) {
+            createNewGroup();
+        } else {
+            alertDialogOpen = true;
+        }
+    }}
+    class="mt-2 w-full">
+    <p class="std-flex">
+        <Icon src={Plus} size="32" class="mr-1" />
+        <span>New Group</span>
+    </p>
+</Button>
+
+<AlertDialog.Root bind:open={alertDialogOpen}>
+    <AlertDialog.Trigger />
+    <AlertDialog.Content>
+        <AlertDialog.Header>
+            <AlertDialog.Title
+                >Are you want to create a new group?</AlertDialog.Title>
+            <AlertDialog.Description>
+                Consider joining an existing group instead.
+            </AlertDialog.Description>
+        </AlertDialog.Header>
+        <AlertDialog.Footer>
+            <AlertDialog.Cancel>Cancel</AlertDialog.Cancel>
+            <Button onclick={createNewGroup}>Create New Group</Button>
+        </AlertDialog.Footer>
+    </AlertDialog.Content>
+</AlertDialog.Root>
+
+<style lang="postcss">
+    .card {
+        @apply flex w-full items-center justify-between border border-t-0 border-input bg-background text-left duration-100 hover:bg-accent hover:text-accent-foreground;
+    }
+</style>
