+++ src/lib/components/JoinDialog.svelte
@@ -1,283 +1,23 @@
 <script lang="ts">
-    import { goto } from "$app/navigation";
-    import {
-        courses,
-        joinDialogOpen,
-        selectedCourse,
-        userGroups
-    } from "$lib/client/state.svelte";
-    import * as AlertDialog from "$lib/components/ui/alert-dialog";
+    import { joinDialogOpen, selectedCourse } from "$lib/client/state.svelte";
     import * as Dialog from "$lib/components/ui/dialog";
-    import { ChevronLeft, Icon, Plus } from "svelte-hero-icons";
-    import { toast } from "svelte-sonner";
-    import Button from "./ui/button/button.svelte";
-    import Input from "./ui/input/input.svelte";
-
-    const normalize = (str: string) => {
-        return str
-            .replace(/[^a-zA-Z0-9]/g, "")
-            .toLowerCase()
-            .trim();
-    };
-
-    let search = $state("");
-    const filteredCourses = $derived(
-        courses.value.filter((course) => {
-            if (userGroups.value.some((x) => x.course.id === course.id)) {
-                return false;
-            }
-
-            const searchNorm = normalize(search);
-
-            if (searchNorm.length < 3) {
-                return false;
-            } else if (searchNorm.length === 3) {
-                return normalize(course.code).includes(searchNorm);
-            } else {
-                return (
-                    normalize(course.code).includes(searchNorm) ||
-                    normalize(course.title).includes(searchNorm)
-                );
-            }
-        })
-    );
-
-    // Clear component data when dialog is closed
-    $effect(() => {
-        if (joinDialogOpen.value === false) {
-            search = "";
-            selectedCourse.value = null;
-        }
-    });
-
-    let alertDialogOpen = $state(false);
-
-    type Group = {
-        groupId: number;
-        groupName: string;
-        members: {
-            netid: string;
-            displayname: string;
-        }[];
-    };
-
-    let isLoading = $state(false);
-    let availableGroups: Group[] = $state([]);
+    import CourseSearch from "./CourseSearch.svelte";
+    import GroupSelection from "./GroupSelection.svelte";
 
     $effect(() => {
-        if (selectedCourse.value !== null) {
-            // Fetch available groups for the selected course
-            isLoading = true;
-
-            try {
-                fetch(`/api/get-groups`, {
-                    method: "POST",
-                    headers: {
-                        "Content-Type": "application/json"
-                    },
-                    body: JSON.stringify({
-                        courseId: selectedCourse.value.id
-                    })
-                })
-                    .then((res) => res.json())
-                    .then((data) => {
-                        availableGroups = data.groups;
-                    });
-            } finally {
-                isLoading = false;
-            }
-        } else {
-            availableGroups = [];
-        }
+        // Clear component data when dialog is closed
+        if (joinDialogOpen.value === false) selectedCourse.value = null;
     });
-
-    const createNewGroup = async () => {
-        if (!selectedCourse.value) {
-            alertDialogOpen = false;
-            return;
-        }
-
-        // Create a new group in the database
-        const res = await fetch("/api/new-group", {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify({
-                courseId: selectedCourse.value.id
-            })
-        });
-        const newGroupId = (await res.json()).group;
-
-        // Redirect to the new group page and cleanup
-        goto(`/group/${newGroupId}`);
-        toast.success(`Created a new group for ${selectedCourse.value.code}!`);
-        alertDialogOpen = false;
-        joinDialogOpen.value = false;
-    };
 </script>
 
 <Dialog.Root bind:open={joinDialogOpen.value}>
     <Dialog.Content class="h-2/3">
         <div class="flex h-full flex-col overflow-hidden">
-            <!-- ! Course Not Selected (search courses) -->
             {#if !selectedCourse.value}
-                <div class="my-4">
-                    <h2 class="text-xl font-semibold">Join Study Group</h2>
-                    <p class="text-sm text-slate-500">
-                        Search for a course to join a group.
-                    </p>
-                </div>
-
-                <Input
-                    bind:value={search}
-                    class="focus-visible:border-slate-300 focus-visible:ring-0"
-                    type="text"
-                    placeholder="Search for courses" />
-
-                <div
-                    class="mt-2 flex flex-1 flex-col overflow-y-auto
-                 {filteredCourses.length > 0 && 'border-t border-input'}">
-                    {#each filteredCourses as course (course.id)}
-                        <button
-                            onclick={() => {
-                                selectedCourse.value = course;
-                            }}
-                            class=" card">
-                            <div class="p-1">
-                                <p class="text-sm font-semibold">
-                                    {course.code}
-                                </p>
-                                <p class="text-xs text-slate-500">
-                                    {course.title}
-                                </p>
-                            </div>
-                        </button>
-                    {:else}
-                        {#if search.length >= 3}
-                            <p class="text-slate-500 text-center py-4">
-                                No courses found.
-                            </p>
-                        {:else}
-                            <p class="text-slate-500 text-center py-4">
-                                Enter at least 3 characters to search.
-                            </p>
-                        {/if}
-                    {/each}
-                </div>
-
-                <!-- ! Course Selected -->
+                <CourseSearch />
             {:else}
-                <div class="my-4">
-                    <h2 class="flex items-center gap-2 text-xl font-semibold">
-                        <span>
-                            <Button
-                                size="icon"
-                                variant="ghost"
-                                onclick={() => {
-                                    selectedCourse.value = null;
-                                }}>
-                                <Icon src={ChevronLeft} size="32" />
-                            </Button>
-                        </span>
-                        <span>
-                            Groups for {selectedCourse.value.code}
-                        </span>
-                    </h2>
-                    <p class="text-sm text-slate-500">
-                        Select a group to join or create a new one.
-                    </p>
-                </div>
-
-                {#if isLoading}
-                    <div class="std-flex mt-4">Loading groups...</div>
-                {:else if availableGroups.length === 0}
-                    <div class="std-flex flex-1 text-slate-500">
-                        No groups found. Create a new one?
-                    </div>
-                {:else}
-                    <div
-                        class="flex flex-1 flex-col overflow-y-auto border-t border-input">
-                        {#each availableGroups as group (group.groupId)}
-                            <button
-                                onclick={async () => {
-                                    await fetch("/api/join-group", {
-                                        method: "POST",
-                                        headers: {
-                                            "Content-Type": "application/json"
-                                        },
-                                        body: JSON.stringify({
-                                            groupId: group.groupId
-                                        })
-                                    });
-                                    joinDialogOpen.value = false;
-                                    goto(`/group/${group.groupId}`);
-                                    toast.success(
-                                        `Joined group ${group.groupName}!`
-                                    );
-                                }}
-                                class="card">
-                                <div class="p-2">
-                                    <p class="text-sm font-semibold">
-                                        Group: {group.groupName}
-                                    </p>
-                                    <p class="text-xs text-slate-500">
-                                        <span class="font-semibold">
-                                            {group.members.length} Members:
-                                        </span>
-                                        <span>
-                                            {group.members
-                                                .map((x) => x.displayname)
-                                                .join(", ")}
-                                        </span>
-                                    </p>
-                                </div>
-                            </button>
-                        {:else}
-                            <div class="flex-1 std-flex text-slate-500">
-                                You are already in all available groups.
-                            </div>
-                        {/each}
-                    </div>
-                {/if}
-                <Button
-                    onclick={() => {
-                        if (availableGroups.length === 0) {
-                            createNewGroup();
-                        } else {
-                            alertDialogOpen = true;
-                        }
-                    }}
-                    class="mt-2 w-full">
-                    <p class="std-flex">
-                        <Icon src={Plus} size="24" />
-                        <span> New Group </span>
-                    </p>
-                </Button>
-                <AlertDialog.Root bind:open={alertDialogOpen}>
-                    <AlertDialog.Trigger />
-                    <AlertDialog.Content>
-                        <AlertDialog.Header>
-                            <AlertDialog.Title
-                                >Are you want to create a new group?</AlertDialog.Title>
-                            <AlertDialog.Description>
-                                Consider joining an existing group instead.
-                            </AlertDialog.Description>
-                        </AlertDialog.Header>
-                        <AlertDialog.Footer>
-                            <AlertDialog.Cancel>Cancel</AlertDialog.Cancel>
-                            <Button onclick={createNewGroup}
-                                >Create New Group</Button>
-                        </AlertDialog.Footer>
-                    </AlertDialog.Content>
-                </AlertDialog.Root>
+                <GroupSelection />
             {/if}
         </div>
     </Dialog.Content>
 </Dialog.Root>
-
-<style lang="postcss">
-    .card {
-        @apply flex w-full items-center justify-between border border-t-0 border-input bg-background text-left duration-100 hover:bg-accent hover:text-accent-foreground;
-    }
-</style>
